amdgpu: Implement multiGPU SVM support v3
With this change, if there are multiple GPU devices,
SVM range and allocation is global to all GPU devices.
This is to meet the OpenCL 2.0 SVM requirement.
This is not a perfect solution. But we have not found
better solution.

Constraints:
  1. Application should initialize all relevant devices
     before allocate SVM address.
  2. If devices do not have similar GPU VM configuration,
     libdrm can disable SVM when new device are initialized.

v3:
a912dddb82071781c21265725556404ee2837e5e
undo reference counter when SVM init fails
869a8f1c8a73aeddaf9439796df89955044553fb
fix to free va for svm at deinit
885418a66eb9922f416ce2551bac9f1297599af9
Mark SVM invalid when deinit SVM in any GPU device.
  To make SVM valid again, application should deinte all
  devices and re-initialize all GPU devices again.

v2:
  1. Put svm_refcount and svm_valid as a field of amdgpu_bo_va_mgr.
  2. Adjust title.

Signed-off-by: Alex Xie <AlexBin.Xie@amd.com>
Reviewed-by: Jammy Zhou <Jammy.Zhou@amd.com>
Reviewed-by: Christian König <christian.koenig@amd.com>
Signed-off-by: Xiaojie Yuan <Xiaojie.Yuan@amd.com>